Squashed 'subtree/Bokeh/' changes from 81d8150..3243657
3243657 leftover merge goofs
8bfc161 Merge branch 'master' into gmap
f85e8ca adding maps 1.  calling create in backbone.js tries to create objects on the server, so I those calls for our client side only ranges 2.  I think the map needs to draw when the div is already on the screen, so I wrap the map calls inside a defer 3.  some of the serialization/deserialization was broken for plot server stuff, so that was fixed 4.  the script_injection of goole maps doesn't really work for plot server, so I just added it to the page dom.
1cad8bd interim work for gmap
58cc25a update bokehjs subtree
55f3875 Squashed 'subtree/bokehjs/' changes from 60f526c..ed6fa08

git-subtree-dir: subtree/Bokeh
git-subtree-split: 3243657c27580dc48a3a39196e889c8acb7b3847